+++ src/Shopware6Package.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Heptacom\HeptaConnect\Package\Shopware6;
+
+use Heptacom\HeptaConnect\Portal\Base\Portal\Contract\PackageContract;
+
+final class Shopware6Package extends PackageContract
+{
+}